wip: Move the serialport code into its own file
